Halting bug fixed
The reason is that when there is no more message and call pumpNetwork
will block process.

So, when finish the loading phase, change to SETUP_GLOBAL_MINMAX
directly. Do not put slaves into NAS_WAITING, this can avoid them
calling pumpNetwork all the time.

This is MPI issue, do not know the root reason, should post question
online and ask.